% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_census_download.R
\name{acs_census}
\alias{acs_census}
\title{\code{acs_census} S3 class}
\usage{
acs_census(dir = "")
}
\arguments{
\item{dir}{A string.}
}
\value{
A \code{acs_census} object.
}
\description{
A \code{acs_census} object is created from a given local dir. This dir will contain
the geodatabase files that we download.
}
\examples{

dir <- system.file("extdata", package = "geogenr")
ac <- acs_census(dir = dir)

}
\seealso{
Other data download functions: 
\code{\link{download_selected_files}()},
\code{\link{get_area_file_names}()},
\code{\link{get_area_groups}()},
\code{\link{get_area_years}()},
\code{\link{get_areas}()},
\code{\link{get_selected_file_names}()},
\code{\link{get_too_heavy_file_names}()},
\code{\link{select_area_files}()},
\code{\link{unzip_files}()}
}
\concept{data download functions}
