% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_metadata_interpret_metadata.R
\name{add_value}
\alias{add_value}
\title{add_value}
\usage{
add_value(mdr, field, value, level = 0)
}
\arguments{
\item{mdr}{A \code{tibble} row.}

\item{field}{A field name to classify the value.}

\item{value}{A value to add.}

\item{level}{Field level to add the value to.}
}
\value{
A \code{tibble} row.
}
\description{
Adds a value to a field at the first free level or at the indicated level.
}
\keyword{internal}
