% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anrc_2021_2022_x01_x07}
\alias{anrc_2021_2022_x01_x07}
\title{"Alaska Native Regional Corporation", 2021, 2022, "X01 Age And Sex", "X07 Migration"}
\format{
A \code{acs_5yr_topic} object.
}
\usage{
anrc_2021_2022_x01_x07
}
\description{
Topics selected for the area and years indicated: "Alaska Native Regional
Corporation", 2021, 2022.
}
\details{
Since no year is indicated in the definition, all available years are considered.
}
\examples{
# Defined by:
dir <- tempdir()
source_dir <- system.file("extdata/acs_5yr", package = "geogenr")
files <- list.files(source_dir, "*.zip", full.names = TRUE)
file.copy(from = files, to = dir, overwrite = TRUE)
ac <- acs_5yr(dir)

files <- ac |>
  unzip_files()

anrc_2021_2022_x01_x07 <- ac |>
  as_acs_5yr_topic(area = "Alaska Native Regional Corporation",
                   topic = c("X01 Age And Sex", "X07 Migration"))

}
\seealso{
Other selection data: 
\code{\link{anrc_2021_x01}}
}
\concept{selection data}
\keyword{datasets}
