% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anrc_2021_x01}
\alias{anrc_2021_x01}
\title{"Alaska Native Regional Corporation", 2021, "X01 Age And Sex"}
\format{
A \code{acs_5yr_topic} object.
}
\usage{
anrc_2021_x01
}
\description{
Topic selected for the area and years indicated: "Alaska Native Regional
Corporation", 2021.
}
\examples{
# Defined by:
dir <- tempdir()
source_dir <- system.file("extdata/acs_5yr", package = "geogenr")
files <- list.files(source_dir, "*.zip", full.names = TRUE)
file.copy(from = files, to = dir, overwrite = TRUE)
ac <- acs_5yr(dir)

files <- ac |>
  unzip_files()

anrc_2021_x01 <- ac |>
  as_acs_5yr_topic("Alaska Native Regional Corporation",
                   2021,
                   "X01 Age And Sex")

}
\seealso{
Other selection data: 
\code{\link{anrc_2021_2022_x01_x07}}
}
\concept{selection data}
\keyword{datasets}
