% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_5yr_geo.R
\name{as_acs_5yr_geo}
\alias{as_acs_5yr_geo}
\alias{as_acs_5yr_geo.acs_5yr_topic}
\title{Get an \code{acs_5yr_geo} object}
\usage{
as_acs_5yr_geo(act)

\method{as_acs_5yr_geo}{acs_5yr_topic}(act)
}
\arguments{
\item{act}{An \code{acs_5yr_topic} object.}
}
\value{
An \code{acs_5yr_geo} object.
}
\description{
Once we have selected the topics that interest us and, possibly also the reports
or subreports, we obtain an \code{acs_5yr_geo} object with which we can represent
or export the geographic layer along with the data of interest more easily.
}
\examples{

act <- anrc_2021_x01 |>
  select_report(report = "B01002-Median Age By Sex")

geo <- act |>
  as_acs_5yr_geo()

}
\seealso{
Other data exploitation and export functions: 
\code{\link{as_GeoPackage}()},
\code{\link{as_flat_table}()},
\code{\link{as_star_database}()},
\code{\link{get_metadata}()},
\code{\link{set_metadata}()}
}
\concept{data exploitation and export functions}
