% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_census.R
\name{as_acs_census_topic}
\alias{as_acs_census_topic}
\alias{as_acs_census_topic.acs_census}
\title{As ACS census topic (report group)}
\usage{
as_acs_census_topic(ac, area, years, topic)

\method{as_acs_census_topic}{acs_census}(ac, area, years = NULL, topic)
}
\arguments{
\item{ac}{A \code{acs_census} object.}

\item{area}{A string, area name.}

\item{years}{A vector, year number.}

\item{topic}{A string, topic name.}
}
\value{
A \code{acs_census_topic} object.
}
\description{
Gets an ACS census topic object (report group) for the given years of the
Demographic and Economic Areas that are downloaded and unzipped, available to
be queried.
}
\examples{

dir <- system.file("extdata", package = "geogenr")
ac <- acs_census(dir)

act <- ac |>
  as_acs_census_topic("State", 2019:2021, "X01 Age And Sex")

act <- ac |>
  as_acs_census_topic("State", topic = "X01 Age And Sex")

}
\seealso{
Other data selection functions: 
\code{\link{get_available_area_topics}()},
\code{\link{get_available_area_years}()},
\code{\link{get_available_areas}()}
}
\concept{data selection functions}
