% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_5yr_rolap.R
\name{as_geomultistar}
\alias{as_geomultistar}
\alias{as_geomultistar.acs_5yr_topic}
\title{As \code{geomultistar::geomultistar} object}
\usage{
as_geomultistar(act, attributes)

\method{as_geomultistar}{acs_5yr_topic}(act, attributes = NULL)
}
\arguments{
\item{act}{An \code{acs_5yr_topic} object.}

\item{attributes}{A string vector.}
}
\value{
A \code{geomultistar} object.
}
\description{
Obtain an \code{geomultistar::geomultistar} object to be able to enrich multidimensional
queries with geographic data.
}
\details{
We can indicate the attributes of the geographic layer to include in the export.
Otherwise, the default attributes are included (not area, perimeter or location
attributes).
}
\examples{

gms <- anrc_2021_x01 |>
  as_geomultistar()

}
\seealso{
Other data exploitation and export functions: 
\code{\link{as_GeoPackage}()},
\code{\link{as_acs_5yr_geo}()},
\code{\link{as_flat_table}()},
\code{\link{as_star_database}()},
\code{\link{get_metadata}()},
\code{\link{set_metadata}()}
}
\concept{data exploitation and export functions}
