% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_metadata_reformat_metadata.R
\name{assign_level}
\alias{assign_level}
\title{assign_level}
\usage{
assign_level(mdr, field)
}
\arguments{
\item{mdr}{A \code{tibble}.}

\item{field}{A field name.}
}
\value{
A \code{tibble}.
}
\description{
Structures the field values in levels so that a value is only in one level.
}
\keyword{internal}
