% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dedata}
\alias{dedata}
\title{Titles and Years of Selected Demographic and Economic Data}
\format{
A vector list.
}
\source{
\url{https://www.census.gov/geographies/mapping-files/time-series/geo/tiger-data.2021.html#list-tab-1656998034}
}
\usage{
dedata
}
\description{
Available selected Demographic and Economic Data from the American Community
Survey (ACS) 5-year estimates data titles and years.
}
\keyword{datasets}
