% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_layer_get_layer.R
\name{define_geomultistar}
\alias{define_geomultistar}
\title{define_geomultistar}
\usage{
define_geomultistar(ft, fact_name, where_names, what_names, place)
}
\arguments{
\item{ft}{A \code{tibble} object.}

\item{fact_name}{A string.}

\item{where_names}{A vector of strings.}

\item{what_names}{A vector of strings.}

\item{place}{A \code{st} object.}
}
\value{
A boolean
}
\description{
Definition of a \code{geomultistar} object from a fact table.
}
\keyword{internal}
