% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_metadata_reformat_metadata.R
\name{delete_empty_columns}
\alias{delete_empty_columns}
\alias{delete_empty_columns.uscb_metadata}
\title{Delete empty columns}
\usage{
delete_empty_columns(um)

\method{delete_empty_columns}{uscb_metadata}(um)
}
\arguments{
\item{um}{A \code{uscb_metadata} object.}
}
\value{
A \code{uscb_metadata} object.
}
\description{
Delete empty columns in metadata table.
}
\keyword{internal}
