% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_5yr_topic.R
\name{delete_last_parenthesis}
\alias{delete_last_parenthesis}
\title{Delete last parenthesis of a string (vectorial)}
\usage{
delete_last_parenthesis(name)
}
\arguments{
\item{name}{A string.}
}
\value{
A string
}
\description{
Delete last parenthesis of a string (vectorial)
}
\keyword{internal}
