% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previous.R
\name{download_area_file}
\alias{download_area_file}
\title{Download area file}
\usage{
download_area_file(file = NULL, out_dir = NULL, delete_zip = FALSE)
}
\value{
A vector, files correctly obtained.
}
\description{
Download the files whose url is indicated, unzip them and, if everything went
well and is indicated in the parameter, delete the downloaded files.
}
\examples{

groups <- get_area_years()

}
\seealso{
Other data selection functions: 
\code{\link{get_area_file_names}()},
\code{\link{get_area_groups}()},
\code{\link{get_area_years}()},
\code{\link{get_areas}()},
\code{\link{get_layer_group_names}()},
\code{\link{get_layer_group}()},
\code{\link{get_layer_names}()},
\code{\link{get_layer}()},
\code{\link{unzip_area_file}()},
\code{\link{uscb_folder}()},
\code{\link{uscb_layer}()}
}
\concept{data selection functions}
