% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_acs_5ye_get_geo.R
\name{download_geodatabases}
\alias{download_geodatabases}
\alias{download_geodatabases.uscb_acs_5ye}
\title{get layer names}
\usage{
download_geodatabases(ua, geodatabase, years, folder = NULL)

\method{download_geodatabases}{uscb_acs_5ye}(ua, geodatabase, years, folder = NULL)
}
\arguments{
\item{ua}{A \code{uscb_acs_5ye} object.}

\item{geodatabase}{A string.}

\item{years}{A vector of years.}

\item{folder}{A string.}
}
\value{
A vector of years.
}
\description{
get layer names.
}
\keyword{internal}
