% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previous.R
\name{get_area_file_names}
\alias{get_area_file_names}
\title{Get area file names}
\usage{
get_area_file_names(area = NULL, years = NULL)
}
\value{
A vector.
}
\description{
Get area file names for the given years. If no year is indicated, all available
ones are obtained.
}
\examples{

groups <- get_area_years()

}
\seealso{
Other data selection functions: 
\code{\link{download_area_file}()},
\code{\link{get_area_groups}()},
\code{\link{get_area_years}()},
\code{\link{get_areas}()},
\code{\link{get_layer_group_names}()},
\code{\link{get_layer_group}()},
\code{\link{get_layer_names}()},
\code{\link{get_layer}()},
\code{\link{unzip_area_file}()},
\code{\link{uscb_folder}()},
\code{\link{uscb_layer}()}
}
\concept{data selection functions}
