% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_census.R
\name{get_available_area_years}
\alias{get_available_area_years}
\alias{get_available_area_years.acs_census}
\title{Get available area years}
\usage{
get_available_area_years(ac, area)

\method{get_available_area_years}{acs_census}(ac, area)
}
\arguments{
\item{ac}{A \code{acs_census} object.}

\item{area}{A string, area name.}
}
\value{
A vector, area years.
}
\description{
Gets the years of the Demographic and Economic Areas that are downloaded and
unzipped, available to be queried.
}
\examples{

dir <- system.file("extdata", package = "geogenr")
ac <- acs_census(dir)

years <- ac |>
  get_available_area_years(area = "State")

}
\seealso{
Other data selection functions: 
\code{\link{as_acs_census_topic}()},
\code{\link{get_available_area_topics}()},
\code{\link{get_available_areas}()}
}
\concept{data selection functions}
