% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_census_selection.R
\name{get_available_areas}
\alias{get_available_areas}
\alias{get_available_areas.acs_census}
\title{Get available area names}
\usage{
get_available_areas(ac)

\method{get_available_areas}{acs_census}(ac)
}
\arguments{
\item{ac}{A \code{acs_census} object.}
}
\value{
A vector, area names.
}
\description{
Gets the names of the Demographic and Economic Areas that are downloaded and
unzipped, available to be queried.
}
\examples{

dir <- system.file("extdata", package = "geogenr")
ac <- acs_census(dir)

areas <- ac |>
  get_available_areas()

}
\seealso{
Other data selection functions: 
\code{\link{as_acs_census_topic}()},
\code{\link{get_available_area_topics}()},
\code{\link{get_available_area_years}()}
}
\concept{data selection functions}
