% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_acs_5ye_get_geo.R
\name{get_available_years}
\alias{get_available_years}
\alias{get_available_years.uscb_acs_5ye}
\title{get layer names}
\usage{
get_available_years(ua)

\method{get_available_years}{uscb_acs_5ye}(ua)
}
\arguments{
\item{ua}{A \code{uscb_acs_5ye} object.}
}
\value{
A vector of years.
}
\description{
get layer names.
}
\keyword{internal}
