% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_acs_5ye_get_geo.R
\name{get_available_years_of_geodatabases}
\alias{get_available_years_of_geodatabases}
\alias{get_available_years_of_geodatabases.uscb_acs_5ye}
\title{get layer names}
\usage{
get_available_years_of_geodatabases(ua, name)

\method{get_available_years_of_geodatabases}{uscb_acs_5ye}(ua, name)
}
\arguments{
\item{ua}{A \code{uscb_acs_5ye} object.}

\item{name}{A string.}
}
\value{
A vector of years.
}
\description{
get layer names.
}
\keyword{internal}
