% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_layer_get_layer.R
\name{get_basic_flat_table}
\alias{get_basic_flat_table}
\alias{get_basic_flat_table.uscb_layer}
\title{get tidy data}
\usage{
get_basic_flat_table(ul, remove_zeros = FALSE)

\method{get_basic_flat_table}{uscb_layer}(ul, remove_zeros = FALSE)
}
\arguments{
\item{ul}{A \code{uscb_layer} object.}

\item{remove_zeros}{A boolean, remove data with zero value.}
}
\value{
A \code{tibble} object.
}
\description{
get tidy data.
}
