% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_folder_get_layer.R
\name{get_common_flat_table}
\alias{get_common_flat_table}
\alias{get_common_flat_table.uscb_folder}
\title{get tidy data}
\usage{
get_common_flat_table(uf, remove_zeros = TRUE, remove_geometry = FALSE)

\method{get_common_flat_table}{uscb_folder}(uf, remove_zeros = TRUE, remove_geometry = TRUE)
}
\arguments{
\item{uf}{A \code{uscb_folder} object.}

\item{remove_zeros}{A boolean, remove data with zero value.}

\item{remove_geometry}{A boolean, remove geometry column.}
}
\value{
A \code{tibble} object.
}
\description{
get tidy data.
}
\keyword{internal}
