% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_folder_get_layer.R
\name{get_common_geomultistar}
\alias{get_common_geomultistar}
\alias{get_common_geomultistar.uscb_folder}
\title{get tidy data}
\usage{
get_common_geomultistar(uf)

\method{get_common_geomultistar}{uscb_folder}(uf)
}
\arguments{
\item{ul}{A \code{uscb_folder} object.}
}
\value{
A \code{geomultistar} object.
}
\description{
get tidy data.
}
\keyword{internal}
