% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_metadata_reformat_metadata.R
\name{get_field_values}
\alias{get_field_values}
\alias{get_field_values.uscb_metadata}
\title{get_field_values}
\usage{
get_field_values(um)

\method{get_field_values}{uscb_metadata}(um)
}
\arguments{
\item{um}{A \code{uscb_metadata} object.}
}
\value{
A \code{uscb_metadata} object.
}
\description{
Get field values.
}
\keyword{internal}
