% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_5yr_common.R
\name{get_file_area}
\alias{get_file_area}
\title{Get file area (with year)}
\usage{
get_file_area(file)
}
\arguments{
\item{file}{A string vector.}
}
\value{
A string vector.
}
\description{
Given a file, get the associated code (in the name), includes the year as the
name of the vector elements..
}
\keyword{internal}
