% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{get_file_code}
\alias{get_file_code}
\title{Get file code (with year)}
\usage{
get_file_code(file)
}
\arguments{
\item{name}{A string vector.}
}
\value{
A string vector.
}
\description{
Given a file, get the associated code (in the name), includes the year as the
name of the vector elements..
}
\keyword{internal}
