% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_layer_get_layer.R
\name{get_flat_table}
\alias{get_flat_table}
\alias{get_flat_table.uscb_layer}
\title{get tidy data}
\usage{
get_flat_table(ul, remove_zeros = FALSE, remove_geometry = FALSE)

\method{get_flat_table}{uscb_layer}(ul, remove_zeros = FALSE, remove_geometry = TRUE)
}
\arguments{
\item{ul}{A \code{uscb_layer} object.}

\item{remove_zeros}{A boolean, remove data with zero value.}

\item{remove_geometry}{A boolean, remove geometry column.}
}
\value{
A \code{tibble} object.
}
\description{
get tidy data.
}
