% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_folder_get_layer.R
\name{get_folder_layer}
\alias{get_folder_layer}
\alias{get_folder_layer.uscb_folder}
\title{get layer}
\usage{
get_folder_layer(uf, layer_name)

\method{get_folder_layer}{uscb_folder}(uf, layer_name)
}
\arguments{
\item{uf}{A \code{uscb_folder} object.}

\item{layer_name}{A layer name.}
}
\value{
A \code{uscb_folder} object.
}
\description{
get layer to interpret variables.
}
