% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_folder_get_layer.R
\name{get_folder_layer_group}
\alias{get_folder_layer_group}
\alias{get_folder_layer_group.uscb_folder}
\title{get layer group}
\usage{
get_folder_layer_group(uf, layer_group_name)

\method{get_folder_layer_group}{uscb_folder}(uf, layer_group_name)
}
\arguments{
\item{uf}{A \code{uscb_folder} object.}

\item{layer_group_name}{A layer name.}
}
\value{
A \code{uscb_folder} object.
}
\description{
get layer group to interpret variables.
}
\keyword{internal}
