% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_folder_get_layer.R
\name{get_folder_layer_names}
\alias{get_folder_layer_names}
\alias{get_folder_layer_names.uscb_folder}
\title{get layer names}
\usage{
get_folder_layer_names(uf)

\method{get_folder_layer_names}{uscb_folder}(uf)
}
\arguments{
\item{uf}{A \code{uscb_folder} object.}
}
\value{
A vector of names.
}
\description{
get layer names.
}
