% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_5yr_topic.R
\name{get_geo_attribute_names}
\alias{get_geo_attribute_names}
\alias{get_geo_attribute_names.acs_5yr_topic}
\title{Get geographical attributes}
\usage{
get_geo_attribute_names(act)

\method{get_geo_attribute_names}{acs_5yr_topic}(act)
}
\arguments{
\item{act}{An \code{acs_5yr_topic} object.}
}
\value{
A vector, geographical attribute names.
}
\description{
Get the names of the geographic layer attributes (except for the geometry field).
}
\examples{

names <- anrc_2021_x01 |>
  get_geo_attribute_names()

}
\seealso{
Other data selection functions: 
\code{\link{as_acs_5yr_topic}()},
\code{\link{as_flat_table}()},
\code{\link{as_star_database}()},
\code{\link{get_available_area_topics}()},
\code{\link{get_available_area_years}()},
\code{\link{get_available_areas}()},
\code{\link{get_geo_layer.acs_5yr_geo}()},
\code{\link{get_names_of_other_topics}()},
\code{\link{get_report_names}()},
\code{\link{get_subreport_names}()},
\code{\link{get_topic_name}()},
\code{\link{select_report}()},
\code{\link{select_subreport}()},
\code{\link{select_topic}()}
}
\concept{data selection functions}
