% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_5yr_geo.R, R/acs_5yr_topic.R
\name{get_geo_layer.acs_5yr_geo}
\alias{get_geo_layer.acs_5yr_geo}
\alias{get_geo_layer}
\alias{get_geo_layer.acs_5yr_topic}
\title{Get geographic layer}
\usage{
\method{get_geo_layer}{acs_5yr_geo}(glc)

get_geo_layer(glc)

\method{get_geo_layer}{acs_5yr_topic}(glc)
}
\arguments{
\item{glc}{A \code{acs_5yr_topic} or \code{acs_5yr_geo} object.}
}
\value{
A \code{sf} object.
}
\description{
Get the geographic layer.
}
\examples{

dir <- tempdir()
source_dir <- system.file("extdata/acs_5yr", package = "geogenr")
files <- list.files(source_dir, "*.zip", full.names = TRUE)
file.copy(from = files, to = dir, overwrite = TRUE)
ac <- acs_5yr(dir)
files <- ac |>
  unzip_files()

act <- ac |>
  as_acs_5yr_topic("Alaska Native Regional Corporation",
                   topic = "X01 Age And Sex")

names <- act |>
  get_geo_layer()

}
\seealso{
Other data selection functions: 
\code{\link{as_GeoPackage}()},
\code{\link{as_acs_5yr_geo}()},
\code{\link{as_acs_5yr_topic}()},
\code{\link{as_flat_table}()},
\code{\link{as_star_database}()},
\code{\link{get_available_area_topics}()},
\code{\link{get_available_area_years}()},
\code{\link{get_available_areas}()},
\code{\link{get_geo_attribute_names}()},
\code{\link{get_metadata}()},
\code{\link{get_names_of_other_topics}()},
\code{\link{get_report_names}()},
\code{\link{get_subreport_names}()},
\code{\link{get_topic_name}()},
\code{\link{select_report}()},
\code{\link{select_subreport}()},
\code{\link{select_topic}()},
\code{\link{set_metadata}()}
}
\concept{data selection functions}
