% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_acs_5ye.R
\name{get_geodatabase_file}
\alias{get_geodatabase_file}
\title{get_geodatabase_file}
\usage{
get_geodatabase_file(folder, extension, names, name, year)
}
\arguments{
\item{folder}{A String}

\item{extension}{A String}

\item{names}{A String}

\item{name}{A String}

\item{year}{An Integer}
}
\value{
A boolean
}
\description{
get_geodatabase_file
}
\keyword{internal}
