% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_layer_get_layer.R
\name{get_geomultistar}
\alias{get_geomultistar}
\alias{get_geomultistar.uscb_layer}
\title{get tidy data}
\usage{
get_geomultistar(ul)

\method{get_geomultistar}{uscb_layer}(ul)
}
\arguments{
\item{ul}{A \code{uscb_layer} object.}
}
\value{
A \code{geomultistar} object.
}
\description{
get tidy data.
}
\keyword{internal}
