% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_layer_get_layer.R
\name{get_layer}
\alias{get_layer}
\alias{get_layer.uscb_layer}
\title{get layer}
\usage{
get_layer(ul, layer_name)

\method{get_layer}{uscb_layer}(ul, layer_name)
}
\arguments{
\item{ul}{A \code{uscb_layer} object.}

\item{layer_name}{A layer name.}
}
\value{
A \code{uscb_layer} object.
}
\description{
get layer to interpret variables.
}
\keyword{internal}
