% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_5yr_topic.R
\name{get_layer_data}
\alias{get_layer_data}
\title{Get layer data}
\usage{
get_layer_data(layer, file)
}
\arguments{
\item{layer}{A string, layer name.}

\item{file}{A string, file name.}
}
\value{
A \code{tibble}, layer data.
}
\description{
Get layer data
}
\keyword{internal}
