% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_layer_get_layer.R
\name{get_layer_group}
\alias{get_layer_group}
\alias{get_layer_group.uscb_layer}
\title{get layer group}
\usage{
get_layer_group(ul, layer_group_name)

\method{get_layer_group}{uscb_layer}(ul, layer_group_name)
}
\arguments{
\item{ul}{A \code{uscb_layer} object.}

\item{layer_group_name}{A layer name.}
}
\value{
A \code{uscb_layer} object.
}
\description{
get layer group to interpret variables.
}
\keyword{internal}
