% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_layer_get_layer.R
\name{get_layer_group_names}
\alias{get_layer_group_names}
\alias{get_layer_group_names.uscb_layer}
\title{get layer group names}
\usage{
get_layer_group_names(ul)

\method{get_layer_group_names}{uscb_layer}(ul)
}
\arguments{
\item{ul}{A \code{uscb_layer} object.}
}
\value{
A vector of names.
}
\description{
get layer group names.
}
