% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_5yr_topic.R
\name{get_other_topics}
\alias{get_other_topics}
\alias{get_other_topics.acs_5yr}
\title{Get other topics (report groups)}
\usage{
get_other_topics(act)

\method{get_other_topics}{acs_5yr}(act)
}
\arguments{
\item{act}{A \code{acs_5yr_topic} object.}
}
\value{
A vector, available report groups.
}
\description{
Get the rest of the available topics with the one that is selected.
}
\examples{

dir <- tempdir()
source_dir <- system.file("extdata/acs_5yr", package = "geogenr")
files <- list.files(source_dir, "*.zip", full.names = TRUE)
file.copy(from = files, to = dir, overwrite = TRUE)
ac <- acs_5yr(dir)
files <- ac |>
  unzip_files()

act <- ac |>
  as_acs_5yr_topic("Alaska Native Regional Corporation",
                   topic = "X01 Age And Sex")

topics <- act |>
  get_other_topics()

}
\seealso{
Other data selection functions: 
\code{\link{as_acs_5yr_topic}()},
\code{\link{get_available_area_topics}()},
\code{\link{get_available_area_years}()},
\code{\link{get_available_areas}()}
}
\concept{data selection functions}
