% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_layer.R
\name{get_place}
\alias{get_place}
\title{get year from filepath}
\usage{
get_place(ul)
}
\arguments{
\item{ul}{A \code{uscb_layer} object.}
}
\value{
A \code{tibble} object.
}
\description{
Get last year from a string.
}
\keyword{internal}
