% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_5yr_topic.R
\name{get_report_names}
\alias{get_report_names}
\alias{get_report_names.acs_5yr_topic}
\title{Get report names}
\usage{
get_report_names(act)

\method{get_report_names}{acs_5yr_topic}(act)
}
\arguments{
\item{act}{A \code{acs_5yr_topic} object.}
}
\value{
A vector, report names.
}
\description{
Get the names of the reports. The report code is included with the name. Each
report can contain multiple subreports.
}
\examples{

dir <- tempdir()
source_dir <- system.file("extdata/acs_5yr", package = "geogenr")
files <- list.files(source_dir, "*.zip", full.names = TRUE)
file.copy(from = files, to = dir, overwrite = TRUE)
ac <- acs_5yr(dir)
files <- ac |>
  unzip_files()

act <- ac |>
  as_acs_5yr_topic("Alaska Native Regional Corporation",
                   topic = "X01 Age And Sex")

reports <- act |>
  get_report_names()

}
\seealso{
Other data selection functions: 
\code{\link{as_acs_5yr_GeoPackage}()},
\code{\link{as_acs_5yr_topic}()},
\code{\link{as_flat_table}()},
\code{\link{as_star_database}()},
\code{\link{get_available_area_topics}()},
\code{\link{get_available_area_years}()},
\code{\link{get_available_areas}()},
\code{\link{get_names_of_other_topics}()},
\code{\link{get_subreport_names}()},
\code{\link{get_topic_name}()},
\code{\link{select_report}()},
\code{\link{select_subreport}()},
\code{\link{select_topic}()}
}
\concept{data selection functions}
