% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_5yr_topic.R
\name{get_subreport_names}
\alias{get_subreport_names}
\alias{get_subreport_names.acs_5yr_topic}
\title{Get subreport names}
\usage{
get_subreport_names(act, report)

\method{get_subreport_names}{acs_5yr_topic}(act, report = NULL)
}
\arguments{
\item{act}{An \code{acs_5yr_topic} object.}

\item{report}{A string, report name.}
}
\value{
A vector, subreport names.
}
\description{
Each topic includes several reports and subreports. Once a topic has been
selected, using this function we obtain the name of the available subreports
of a report. If no report is indicated, all subreports of the topic are
obtained.
}
\examples{

reports <- anrc_2021_x01 |>
  get_subreport_names(report = "B01002-Median Age By Sex")

}
\seealso{
Other data selection functions: 
\code{\link{as_acs_5yr_topic}()},
\code{\link{get_available_area_topics}()},
\code{\link{get_available_area_years}()},
\code{\link{get_available_areas}()},
\code{\link{get_geo_attribute_names}()},
\code{\link{get_geo_layer.acs_5yr_geo}()},
\code{\link{get_names_of_other_topics}()},
\code{\link{get_report_names}()},
\code{\link{get_topic_name}()},
\code{\link{select_report}()},
\code{\link{select_subreport}()},
\code{\link{select_topic}()}
}
\concept{data selection functions}
