% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_layer_get_layer.R
\name{get_tidy_data}
\alias{get_tidy_data}
\alias{get_tidy_data.uscb_layer}
\title{get tidy data}
\usage{
get_tidy_data(ul, remove_zeros = TRUE, remove_geometry = FALSE)

\method{get_tidy_data}{uscb_layer}(ul, remove_zeros = TRUE, remove_geometry = FALSE)
}
\arguments{
\item{ul}{A \code{uscb_layer} object.}

\item{remove_zeros}{A boolean, remove data with zero value.}

\item{remove_geometry}{A boolean, remove geometry column.}
}
\value{
A \code{tibble} object.
}
\description{
get tidy data.
}
\keyword{internal}
