% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{get_too_heavy_file_names}
\alias{get_too_heavy_file_names}
\alias{get_too_heavy_file_names.acs_census}
\title{Get too heavy file names}
\usage{
get_too_heavy_file_names()

get_too_heavy_file_names.acs_census(ac)
}
\arguments{
\item{ac}{A \code{acs_census} object.}
}
\value{
A vector, too heavy file names.
}
\description{
Gets the names of the files that are too heavy to be download with the available
function.
}
\examples{

dir <- system.file("extdata", package = "geogenr")
ac <- acs_census(dir)

groups <- ac |>
  get_too_heavy_file_names()

}
\seealso{
Other data download functions: 
\code{\link{acs_census}()},
\code{\link{download_area_file}()},
\code{\link{get_area_file_names}()},
\code{\link{get_area_groups}()},
\code{\link{get_area_years}()},
\code{\link{get_areas}()},
\code{\link{get_downloaded_areas}()},
\code{\link{select_area_files}()},
\code{\link{unzip_area_file}()}
}
\concept{data download functions}
