% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_5yr_topic.R
\name{get_topic_data}
\alias{get_topic_data}
\title{Select topic (report group)}
\usage{
get_topic_data(act)
}
\arguments{
\item{act}{An \code{acs_5yr_topic} object.}
}
\value{
An \code{acs_5yr_topic} object.
}
\description{
Select the given topic with. If no topic is given, the first one that appears
in the files is taken.
}
\keyword{internal}
