% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_layer.R
\name{get_year_from_filepath}
\alias{get_year_from_filepath}
\title{Get year from filepath}
\usage{
get_year_from_filepath(filepath)
}
\arguments{
\item{mdr}{A string.}
}
\value{
A string.
}
\description{
Get last year from a string.
}
\keyword{internal}
