% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_acs_5ye_get_geo.R
\name{get_years_available}
\alias{get_years_available}
\alias{get_years_available.uscb_acs_5ye}
\title{get layer names}
\usage{
get_years_available(ua)

\method{get_years_available}{uscb_acs_5ye}(ua)
}
\arguments{
\item{ua}{A \code{uscb_acs_5ye} object.}
}
\value{
A vector of names.
}
\description{
get layer names.
}
\keyword{internal}
