% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiger_acs.R, R/uscb_metadata.R
\name{\%>\%}
\alias{\%>\%}
\alias{new_tiger_acs}
\alias{new_uscb_metadata}
\title{\code{tiger_acs} S3 class}
\usage{
new_tiger_acs(filepath = NULL, code = NULL)

new_uscb_metadata(filepath = NULL, code = NULL)
}
\arguments{
\item{filepath}{A string, path to gbd file.}

\item{code}{A string.}
}
\value{
A \code{tiger_acs} object.

A \code{uscb_metadata} object.
}
\description{
Internal low-level constructor that creates new objects with the correct
structure.

Internal low-level constructor that creates new objects with the correct
structure.
}
\keyword{internal}
