% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_metadata_interpret_metadata.R
\name{interpret_all}
\alias{interpret_all}
\title{interpret_all}
\usage{
interpret_all(mdr, val, value, interpret, field_values, other_field)
}
\arguments{
\item{mdr}{A \code{tibble} row.}

\item{val}{A transformed value.}

\item{value}{A value to classify.}

\item{interpret}{Vector of functions to consider.}

\item{field_values}{A data frame that stores associations between fields and
values.}

\item{other_field}{Included field that can contain the value "other".}
}
\value{
A result structure.
}
\description{
Apply each of the available functions to the value to classify it.
}
\keyword{internal}
