% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_metadata_interpret_metadata.R
\name{interpret_as}
\alias{interpret_as}
\title{interpret_as}
\usage{
interpret_as(mdr, field, val, value, val_set = NULL, field_values = NULL)
}
\arguments{
\item{mdr}{A \code{tibble} row.}

\item{field}{A field name to classify the value.}

\item{val}{A transformed value.}

\item{value}{A value.}

\item{val_set}{Set of possible values for the field.}

\item{field_values}{A data frame to store associations between fields and
values.}
}
\value{
A result structure.
}
\description{
Classifies a value within a field if it is one of the field's possible
values.
}
\keyword{internal}
