% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_metadata_interpret_metadata.R
\name{interpret_as_housing_computer_and_internet_use}
\alias{interpret_as_housing_computer_and_internet_use}
\title{interpret_as_housing_computer_and_internet_use}
\usage{
interpret_as_housing_computer_and_internet_use(mdr, val, value)
}
\arguments{
\item{mdr}{A \code{tibble} row.}

\item{val}{A transformed value.}

\item{value}{A value.}
}
\value{
A result structure.
}
\description{
Classifies the value in a field if it is one of the possible values
considered for that field.
}
\keyword{internal}
