% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_metadata_interpret_metadata.R
\name{interpret_as_social_grandparents_as_caregivers}
\alias{interpret_as_social_grandparents_as_caregivers}
\title{interpret_as_social_grandparents_as_caregivers}
\usage{
interpret_as_social_grandparents_as_caregivers(mdr, val, value)
}
\arguments{
\item{mdr}{A \code{tibble} row.}

\item{val}{A transformed value.}

\item{value}{A value.}
}
\value{
A result structure.
}
\description{
Classifies the value in a field if it is one of the possible values
considered for that field.
}
\keyword{internal}
