% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_metadata_interpret_metadata.R
\name{interpret_as_social_language_spoken_at_home}
\alias{interpret_as_social_language_spoken_at_home}
\title{interpret_as_social_language_spoken_at_home}
\usage{
interpret_as_social_language_spoken_at_home(
  mdr,
  val,
  value,
  field_values = NULL
)
}
\arguments{
\item{mdr}{A \code{tibble} row.}

\item{val}{A transformed value.}

\item{value}{A value.}

\item{field_values}{A data frame to store associations between fields and
values.}
}
\value{
A result structure.
}
\description{
Classifies the value in a field if it is one of the possible values
considered for that field.
}
\keyword{internal}
