% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_metadata_interpret_metadata.R
\name{interpret_code}
\alias{interpret_code}
\title{interpret code}
\usage{
interpret_code(mdr)
}
\arguments{
\item{mdr}{A \code{tibble} row.}
}
\value{
A \code{tibble} row.
}
\description{
Interprets the values of the code included in the metadata for a row.
}
\keyword{internal}
