% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_metadata_interpret_metadata.R
\name{interpret_metadata}
\alias{interpret_metadata}
\alias{interpret_metadata.uscb_metadata}
\title{Interpret metadata}
\usage{
interpret_metadata(ta)

\method{interpret_metadata}{uscb_metadata}(ta)
}
\arguments{
\item{ta}{A string.}
}
\value{
A \code{uscb_metadata} object.
}
\description{
Interpret metadata
}
\keyword{internal}
