% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_metadata_interpret_metadata.R
\name{interpret_values}
\alias{interpret_values}
\title{interpret code}
\usage{
interpret_values(mdr, values)
}
\arguments{
\item{mdr}{A \code{tibble} row.}

\item{values}{A vector of values}
}
\value{
A \code{tibble} row.
}
\description{
interpret code
}
\keyword{internal}
