% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_census.R, R/acs_census_common.R
\name{name_to_title}
\alias{name_to_title}
\title{Name to title}
\usage{
name_to_title(name)

name_to_title(name)
}
\arguments{
\item{name}{A string.}
}
\value{
A string

A string
}
\description{
Name to title

Name to title
}
\keyword{internal}
