% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_5yr_selection.R
\name{new_acs_5yr_topic}
\alias{new_acs_5yr_topic}
\title{New acs_5yr_topic object}
\usage{
new_acs_5yr_topic(ac, area, years = NULL, topic = NULL)
}
\arguments{
\item{ac}{A \code{acs_5yr} object.}

\item{area}{A string, area name.}

\item{years}{A vector, year number.}

\item{topic}{A string, topic name.}
}
\value{
A \code{acs_5yr_topic} object.
}
\description{
New acs_5yr_topic object
}
\keyword{internal}
