% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_acs_5ye.R
\name{new_uscb_acs_5ye}
\alias{new_uscb_acs_5ye}
\title{\code{uscb_acs_5ye} S3 class}
\usage{
new_uscb_acs_5ye(folder = "")
}
\arguments{
\item{folder}{A string.}
}
\value{
A \code{uscb_acs_5ye} object.
}
\description{
Internal low-level constructor that creates new objects with the correct
structure.
}
\keyword{internal}
