% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_folder.R
\name{new_uscb_folder}
\alias{new_uscb_folder}
\title{\code{uscb_folder} S3 class}
\usage{
new_uscb_folder(ul, year = NULL)
}
\arguments{
\item{ul}{A \code{uscb_folder} object.}

\item{year}{A vector of years.}
}
\value{
A \code{uscb_folder} object.
}
\description{
Internal low-level constructor that creates new objects with the correct
structure.
}
\keyword{internal}
