% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_metadata_reformat_metadata.R
\name{reformat_metadata}
\alias{reformat_metadata}
\alias{reformat_metadata.uscb_metadata}
\title{Reformat metadata}
\usage{
reformat_metadata(um)

\method{reformat_metadata}{uscb_metadata}(um)
}
\arguments{
\item{um}{A \code{uscb_metadata} object.}
}
\value{
A \code{uscb_metadata} object.
}
\description{
Reformat the metadata to distribute it in columns according to its topic.
}
\keyword{internal}
