% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_metadata_reformat_metadata.R
\name{replace_numbers}
\alias{replace_numbers}
\title{Replace numbers}
\usage{
replace_numbers(docs)
}
\arguments{
\item{docs}{A character vector.}
}
\value{
A character vector.
}
\description{
Replace numbers by N.
}
\keyword{internal}
