% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_folder_get_layer.R
\name{same_layer_group_columns}
\alias{same_layer_group_columns}
\title{same_layer_group_columns}
\usage{
same_layer_group_columns(initial, rest)
}
\arguments{
\item{initial}{A vector of strings.}

\item{rest}{A vector of strings.}
}
\value{
A boolean
}
\description{
Check if two layer groups have the same columns.
}
\keyword{internal}
