% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_metadata_reformat_metadata.R
\name{scroll_level}
\alias{scroll_level}
\title{scroll_level}
\usage{
scroll_level(mdr, fields, field_index, values_indices, index_limit)
}
\arguments{
\item{mdr}{A \code{tibble}.}

\item{fields}{Vector with the name of the field and its levels.}

\item{field_index}{Index of the vector of names to consider.}

\item{values_indices}{Indices of the \code{tibble} positions to move.}

\item{index_limit}{Level limit for the name index.}
}
\value{
A \code{tibble}.
}
\description{
Recursive function. Shifts the required values to be able to move the set of
values to the next level.
}
\keyword{internal}
