% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_5yr_geo.R
\name{set_metadata}
\alias{set_metadata}
\alias{set_metadata.acs_5yr_geo}
\title{Set metadata layer}
\usage{
set_metadata(geo, metadata)

\method{set_metadata}{acs_5yr_geo}(geo, metadata)
}
\arguments{
\item{geo}{A \code{acs_5yr_geo} object.}

\item{metadata}{A \code{tibble} object.}
}
\value{
A \code{sf} object.
}
\description{
Set the metadata layer.
}
\examples{

dir <- tempdir()
source_dir <- system.file("extdata/acs_5yr", package = "geogenr")
files <- list.files(source_dir, "*.zip", full.names = TRUE)
file.copy(from = files, to = dir, overwrite = TRUE)
ac <- acs_5yr(dir)
files <- ac |>
  unzip_files()

act <- ac |>
  as_acs_5yr_topic("Alaska Native Regional Corporation",
                   topic = "X01 Age And Sex")

metadata <- act |>
  get_metadata()

act2 <- act |>
  set_metadata(metadata)

}
\seealso{
Other data selection functions: 
\code{\link{as_GeoPackage}()},
\code{\link{as_acs_5yr_geo}()},
\code{\link{as_acs_5yr_topic}()},
\code{\link{as_flat_table}()},
\code{\link{as_star_database}()},
\code{\link{get_available_area_topics}()},
\code{\link{get_available_area_years}()},
\code{\link{get_available_areas}()},
\code{\link{get_geo_attribute_names}()},
\code{\link{get_geo_layer.acs_5yr_geo}()},
\code{\link{get_metadata}()},
\code{\link{get_names_of_other_topics}()},
\code{\link{get_report_names}()},
\code{\link{get_subreport_names}()},
\code{\link{get_topic_name}()},
\code{\link{select_report}()},
\code{\link{select_subreport}()},
\code{\link{select_topic}()}
}
\concept{data selection functions}
