% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_metadata_reformat_metadata.R
\name{show_fields}
\alias{show_fields}
\alias{show_fields.uscb_metadata}
\title{Show fields}
\usage{
show_fields(um)

\method{show_fields}{uscb_metadata}(um)
}
\arguments{
\item{um}{A \code{uscb_metadata} object.}
}
\value{
A \code{uscb_metadata} object.
}
\description{
Show fields in metadata table.
}
\keyword{internal}
