% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_metadata_reformat_metadata.R
\name{standardize_text}
\alias{standardize_text}
\title{standardize text}
\usage{
standardize_text(docs)
}
\arguments{
\item{mdr}{A character vector.}
}
\value{
A character vector.
}
\description{
standardize text making transformations.
}
\keyword{internal}
