% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_metadata_reformat_metadata.R
\name{standardize_text2}
\alias{standardize_text2}
\title{Standardize text 2}
\usage{
standardize_text2(docs)
}
\arguments{
\item{docs}{A character vector.}
}
\value{
A character vector.
}
\description{
Standardize text 2 making transformations.
}
\keyword{internal}
