% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tiger_acs_metadata}
\alias{tiger_acs_metadata}
\title{tiger_acs_metadata}
\format{
A \code{tiger_acs}.
}
\source{
\url{https://CRAN.R-project.org/package=starschemar}
}
\usage{
tiger_acs_metadata
}
\description{
tiger_acs_metadata
}
\keyword{datasets}
