% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_5yr_topic.R
\name{transform_layer}
\alias{transform_layer}
\title{Transform layer according to metadata}
\usage{
transform_layer(layer, metadata)
}
\arguments{
\item{layer}{A \code{tibble}, layer data.}

\item{metadata}{A \code{tibble}, layer metadata.}
}
\value{
A vector
}
\description{
Transform layer according to metadata
}
\keyword{internal}
