% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_5yr_topic.R
\name{transform_metadata_basic}
\alias{transform_metadata_basic}
\title{Transform metadata layer}
\usage{
transform_metadata_basic(metadata)
}
\arguments{
\item{metadata}{A vector}
}
\value{
A vector
}
\description{
Transform metadata layer
}
\keyword{internal}
