% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previous.R
\name{unzip_area_file}
\alias{unzip_area_file}
\title{Unzip area file}
\usage{
unzip_area_file(file, out_dir = NULL, only_show_files = FALSE)
}
\arguments{
\item{file}{A string or string vector.}

\item{out_dir}{A string or string vector, output folder.}

\item{only_show_files}{A boolean, only show the files that would be unzipped,
and the destination folders, not unzip them.}
}
\value{
A vector of strings, name of the processed files.
}
\description{
Given a vector of compressed file names or the name of a folder containing
compressed files, unzip the files to the given output folder.
}
\examples{

f <- system.file("extdata", package = "satres")
r <- sat_untarzip(f, only_show_files = TRUE)

f1 <- system.file("extdata", "satres.zip", package = "satres")
f2 <- system.file("extdata", "satres.tar", package = "satres")
r <- sat_untarzip(c(f1, f2), only_show_files = TRUE)

}
\seealso{
Other data selection functions: 
\code{\link{download_area_file}()},
\code{\link{get_area_file_names}()},
\code{\link{get_area_groups}()},
\code{\link{get_area_years}()},
\code{\link{get_areas}()},
\code{\link{get_layer_group_names}()},
\code{\link{get_layer_group}()},
\code{\link{get_layer_names}()},
\code{\link{get_layer}()},
\code{\link{uscb_folder}()},
\code{\link{uscb_layer}()}
}
\concept{data selection functions}
