% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_acs_5ye.R
\name{url_file_exists}
\alias{url_file_exists}
\title{url_file_exists}
\usage{
url_file_exists(url)
}
\arguments{
\item{url}{A string.}
}
\value{
A boolean
}
\description{
https://stackoverflow.com/questions/60318926/how-to-check-if-file-exists-in-the-url-before-use-download-file-in-r
}
\keyword{internal}
