% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_acs_5ye.R
\name{uscb_acs_5ye}
\alias{uscb_acs_5ye}
\title{\code{uscb_acs_5ye} S3 class}
\usage{
uscb_acs_5ye(folder = "")
}
\arguments{
\item{folder}{A string.}
}
\value{
A \code{uscb_acs_5ye} object.
}
\description{
A \code{uscb_acs_5ye} object is created from a given
}
