% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uscb_acs_metadata}
\alias{uscb_acs_metadata}
\title{\code{uscb_acs_metadata}}
\format{
A \code{uscb_metadata}.
}
\source{
\url{https://www.census.gov/programs-surveys/acs/guidance/subjects.html}
}
\usage{
uscb_acs_metadata
}
\description{
Metadata according to subjects included in the American Community Survey
(ACS) from the United States Census Bureau (USCB).
}
\details{
They are obtained from the variables defined in the reports, classifying the
concepts according to these subjects.
}
\keyword{datasets}
