% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_folder.R
\name{uscb_folder}
\alias{uscb_folder}
\title{\code{uscb_folder} S3 class}
\usage{
uscb_folder(ul, year = NULL)
}
\arguments{
\item{ul}{A \code{uscb_folder} object.}

\item{year}{A vector of years.}
}
\value{
A \code{uscb_folder} object.
}
\description{
A \code{uscb_folder} object is created from a given
}
