% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_layer.R
\name{uscb_layer}
\alias{uscb_layer}
\title{\code{uscb_layer} S3 class}
\usage{
uscb_layer(filepath = NULL, metadata = NULL)
}
\arguments{
\item{filepath}{A string, path to gbd file.}

\item{metadata}{A metadata object.}
}
\value{
A \code{uscb_layer} object.
}
\description{
A \code{uscb_layer} object is created from a given
}
