% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_metadata.R
\name{uscb_metadata}
\alias{uscb_metadata}
\title{\code{uscb_metadata} S3 class}
\usage{
uscb_metadata(filepath = NULL, code = NULL, group_code = NULL)
}
\arguments{
\item{filepath}{A string, path to gbd file.}

\item{code}{A string.}

\item{group_code}{A string.}
}
\value{
A \code{uscb_metadata} object.
}
\description{
A \code{uscb_metadata} object is created from a given
}
