% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_5yr_selection.R
\name{as_acs_5yr_topic}
\alias{as_acs_5yr_topic}
\alias{as_acs_5yr_topic.acs_5yr}
\title{As ACS census topic (report group)}
\usage{
as_acs_5yr_topic(ac, area, years, topic)

\method{as_acs_5yr_topic}{acs_5yr}(ac, area, years = NULL, topic = NULL)
}
\arguments{
\item{ac}{A \code{acs_5yr} object.}

\item{area}{A string, area name.}

\item{years}{A vector, year number.}

\item{topic}{A string, topic name.}
}
\value{
A \code{acs_5yr_topic} object.
}
\description{
Gets an ACS census topic object (report group) for the given years of the
Demographic and Economic Areas that are downloaded and unzipped, available to
be queried.
}
\details{
If no year is indicated, all available years are taken. If no topic is given,
the first one that appears in the files is taken.
}
\examples{

dir <- tempdir()
source_dir <- system.file("extdata/acs_5yr", package = "geogenr")
files <- list.files(source_dir, "*.zip", full.names = TRUE)
file.copy(from = files, to = dir, overwrite = TRUE)
ac <- acs_5yr(dir)

files <- ac |>
  unzip_files()

act <- ac |>
  as_acs_5yr_topic("Alaska Native Regional Corporation",
                   2021,
                   "X01 Age And Sex")

act <- ac |>
  as_acs_5yr_topic("Alaska Native Regional Corporation",
                   topic = "X01 Age And Sex")

}
\seealso{
Other data selection functions: 
\code{\link{get_available_area_topics}()},
\code{\link{get_available_area_years}()},
\code{\link{get_available_areas}()},
\code{\link{get_other_topics}()}
}
\concept{data selection functions}
