% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_5yr_geo.R
\name{as_GeoPackage}
\alias{as_GeoPackage}
\alias{as_GeoPackage.acs_5yr_geo}
\title{Save as GeoPackage}
\usage{
as_GeoPackage(geo, dir, name)

\method{as_GeoPackage}{acs_5yr_geo}(geo, dir = NULL, name = NULL)
}
\arguments{
\item{geo}{A \code{acs_5yr_geo} object.}

\item{dir}{A string.}

\item{name}{A string, file name.}
}
\value{
A vector, geographical attribute names.
}
\description{
Get the names of the geographic layer attributes (except for the geometry field).
}
\examples{

dir <- tempdir()
source_dir <- system.file("extdata/acs_5yr", package = "geogenr")
files <- list.files(source_dir, "*.zip", full.names = TRUE)
file.copy(from = files, to = dir, overwrite = TRUE)
ac <- acs_5yr(dir)
files <- ac |>
  unzip_files()

act <- ac |>
  as_acs_5yr_topic("Alaska Native Regional Corporation",
                   topic = "X01 Age And Sex")

names <- act |>
  as_GeoPackage()

}
\seealso{
Other data selection functions: 
\code{\link{as_acs_5yr_geo}()},
\code{\link{as_acs_5yr_topic}()},
\code{\link{as_flat_table}()},
\code{\link{as_star_database}()},
\code{\link{get_available_area_topics}()},
\code{\link{get_available_area_years}()},
\code{\link{get_available_areas}()},
\code{\link{get_geo_attribute_names}()},
\code{\link{get_geo_layer.acs_5yr_geo}()},
\code{\link{get_names_of_other_topics}()},
\code{\link{get_report_names}()},
\code{\link{get_subreport_names}()},
\code{\link{get_topic_name}()},
\code{\link{select_report}()},
\code{\link{select_subreport}()},
\code{\link{select_topic}()}
}
\concept{data selection functions}
