% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{get_area_file_names}
\alias{get_area_file_names}
\title{Get area file names}
\usage{
get_area_file_names(area = NULL, years = NULL)
}
\arguments{
\item{area}{A string, area name.}

\item{years}{A vector, year number.}
}
\value{
A vector, file urls.
}
\description{
Get area url file names for the given years. If no year is indicated, all
available ones are obtained.
}
\examples{

url <- get_area_file_names("State", 2019:2021)

url <- get_area_file_names("State")

}
\seealso{
Other data download functions: 
\code{\link{download_area_file}()},
\code{\link{get_area_groups}()},
\code{\link{get_area_years}()},
\code{\link{get_areas}()},
\code{\link{unzip_area_file}()}
}
\concept{data download functions}
