% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previous.R
\name{get_area_groups}
\alias{get_area_groups}
\title{Get area groups}
\usage{
get_area_groups()
}
\value{
A vector.
}
\description{
Gets the names of the Demographic and Economic area groups where data is available.
}
\examples{

groups <- get_area_groups()

}
\seealso{
Other data selection functions: 
\code{\link{download_area_file}()},
\code{\link{get_area_file_names}()},
\code{\link{get_area_years}()},
\code{\link{get_areas}()},
\code{\link{get_layer_group_names}()},
\code{\link{get_layer_group}()},
\code{\link{get_layer_names}()},
\code{\link{get_layer}()},
\code{\link{unzip_area_file}()},
\code{\link{uscb_folder}()},
\code{\link{uscb_layer}()}
}
\concept{data selection functions}
