% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{get_area_groups}
\alias{get_area_groups}
\title{Get area groups}
\usage{
get_area_groups()
}
\value{
A vector, area group names.
}
\description{
Gets the names of the Demographic and Economic Area Groups where data is available.
}
\examples{

groups <- get_area_groups()

}
\seealso{
Other data download functions: 
\code{\link{download_area_file}()},
\code{\link{get_area_file_names}()},
\code{\link{get_area_years}()},
\code{\link{get_areas}()},
\code{\link{unzip_area_file}()}
}
\concept{data download functions}
