% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previous.R
\name{get_area_years}
\alias{get_area_years}
\title{Get available area years}
\usage{
get_area_years(area = NULL)
}
\value{
A vector.
}
\description{
Get the years for which data has been found to be available for an area.
}
\examples{

groups <- get_area_years()

}
\seealso{
Other data selection functions: 
\code{\link{download_area_file}()},
\code{\link{get_area_file_names}()},
\code{\link{get_area_groups}()},
\code{\link{get_areas}()},
\code{\link{get_layer_group_names}()},
\code{\link{get_layer_group}()},
\code{\link{get_layer_names}()},
\code{\link{get_layer}()},
\code{\link{unzip_area_file}()},
\code{\link{uscb_folder}()},
\code{\link{uscb_layer}()}
}
\concept{data selection functions}
