% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_obtaining.R
\name{get_area_years}
\alias{get_area_years}
\title{Get available area years}
\usage{
get_area_years(area = NULL)
}
\arguments{
\item{area}{A string, area name.}
}
\value{
A vector, area years.
}
\description{
Get the years for which data has been found to be available for an area.
}
\examples{

years <- get_area_years("State")

}
\seealso{
Other data download functions: 
\code{\link{download_area_file}()},
\code{\link{get_area_file_names}()},
\code{\link{get_area_groups}()},
\code{\link{get_areas}()},
\code{\link{unzip_area_file}()}
}
\concept{data download functions}
