% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previous.R
\name{get_areas}
\alias{get_areas}
\title{Get area names of a group}
\usage{
get_areas(group = NULL)
}
\value{
A vector.
}
\description{
Gets the names of the Demographic and Economic areas of a group or set of groups.
}
\details{
If no group is indicated, all available areas are obtained.
}
\examples{

areas <- get_areas(group = "Statistical Areas")

}
\seealso{
Other data selection functions: 
\code{\link{download_area_file}()},
\code{\link{get_area_file_names}()},
\code{\link{get_area_groups}()},
\code{\link{get_area_years}()},
\code{\link{get_layer_group_names}()},
\code{\link{get_layer_group}()},
\code{\link{get_layer_names}()},
\code{\link{get_layer}()},
\code{\link{unzip_area_file}()},
\code{\link{uscb_folder}()},
\code{\link{uscb_layer}()}
}
\concept{data selection functions}
