% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{get_areas}
\alias{get_areas}
\title{Get area names of a group}
\usage{
get_areas(group = NULL)
}
\arguments{
\item{group}{A string, area group name.}
}
\value{
A vector, area names.
}
\description{
Gets the names of the Demographic and Economic Areas of a group or set of groups.
}
\details{
If no group is indicated, all available areas are obtained.
}
\examples{

areas <- get_areas(group = "Statistical Areas")

}
\seealso{
Other data download functions: 
\code{\link{download_area_file}()},
\code{\link{get_area_file_names}()},
\code{\link{get_area_groups}()},
\code{\link{get_area_years}()},
\code{\link{unzip_area_file}()}
}
\concept{data download functions}
