% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_census.R, R/acs_census_selection.R
\name{get_available_area_topics}
\alias{get_available_area_topics}
\alias{get_available_area_topics.acs_census}
\title{Get available area topics (report groups)}
\usage{
get_available_area_topics(ac, area, years)

\method{get_available_area_topics}{acs_census}(ac, area, years = NULL)

get_available_area_topics(ac, area, years)

\method{get_available_area_topics}{acs_census}(ac, area, years = NULL)
}
\arguments{
\item{ac}{A \code{acs_census} object.}

\item{area}{A string, area name.}

\item{years}{A vector, year number.}
}
\value{
A vector, available report groups.

A vector, available report groups.
}
\description{
Gets the topics (report groups) for the given years of the Demographic and
Economic Areas that are downloaded and unzipped, available to be queried.

Gets the topics (report groups) for the given years of the Demographic and
Economic Areas that are downloaded and unzipped, available to be queried.
}
\examples{

dir <- system.file("extdata", package = "geogenr")
ac <- acs_census(dir)

topics <- ac |>
  get_available_area_topics("State", 2019:2021)

topics <- ac |>
  get_available_area_topics("State")


dir <- system.file("extdata", package = "geogenr")
ac <- acs_census(dir)

topics <- ac |>
  get_available_area_topics("State", 2019:2021)

topics <- ac |>
  get_available_area_topics("State")

}
\seealso{
Other data selection functions: 
\code{\link{as_acs_census_topic}()},
\code{\link{get_available_area_years}()},
\code{\link{get_available_areas}()}

Other data selection functions: 
\code{\link{as_acs_census_topic}()},
\code{\link{get_available_area_years}()},
\code{\link{get_available_areas}()}
}
\concept{data selection functions}
