% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_5yr_selection.R
\name{get_available_areas}
\alias{get_available_areas}
\alias{get_available_areas.acs_5yr}
\title{Get available area names}
\usage{
get_available_areas(ac)

\method{get_available_areas}{acs_5yr}(ac)
}
\arguments{
\item{ac}{A \code{acs_5yr} object.}
}
\value{
A vector, area names.
}
\description{
Gets the names of the Demographic and Economic Areas that are downloaded and
unzipped, available to be queried.
}
\examples{

dir <- tempdir()
source_dir <- system.file("extdata/acs_5yr", package = "geogenr")
files <- list.files(source_dir, "*.zip", full.names = TRUE)
file.copy(from = files, to = dir, overwrite = TRUE)
ac <- acs_5yr(dir)

files <- ac |>
  unzip_files()

areas <- ac |>
  get_available_areas()

}
\seealso{
Other data selection functions: 
\code{\link{as_acs_5yr_topic}()},
\code{\link{get_available_area_topics}()},
\code{\link{get_available_area_years}()},
\code{\link{get_other_topics}()}
}
\concept{data selection functions}
