% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_acs_5ye_get_geo.R
\name{get_available_years_downloaded}
\alias{get_available_years_downloaded}
\alias{get_available_years_downloaded.uscb_acs_5ye}
\title{get layer names}
\usage{
get_available_years_downloaded(ua, geodatabase, folder = NULL)

\method{get_available_years_downloaded}{uscb_acs_5ye}(ua, geodatabase, folder = NULL)
}
\arguments{
\item{ua}{A \code{uscb_acs_5ye} object.}

\item{geodatabase}{A string.}

\item{folder}{A string.}
}
\value{
A vector of years.
}
\description{
get layer names.
}
\keyword{internal}
