% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_acs_5ye_get_geo.R
\name{get_available_years_in_the_web}
\alias{get_available_years_in_the_web}
\alias{get_available_years_in_the_web.uscb_acs_5ye}
\title{get layer names}
\usage{
get_available_years_in_the_web(ua, geodatabase)

\method{get_available_years_in_the_web}{uscb_acs_5ye}(ua, geodatabase)
}
\arguments{
\item{ua}{A \code{uscb_acs_5ye} object.}

\item{geodatabase}{A string.}
}
\value{
A vector of years.
}
\description{
get layer names.
}
\keyword{internal}
