% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_layer_get_layer.R
\name{get_basic_flat_table}
\alias{get_basic_flat_table}
\alias{get_basic_flat_table.uscb_layer}
\title{Get basic flat table}
\usage{
get_basic_flat_table(ul, remove_zeros = FALSE)

\method{get_basic_flat_table}{uscb_layer}(ul, remove_zeros = FALSE)
}
\arguments{
\item{ul}{A \code{uscb_layer} object.}

\item{remove_zeros}{A boolean, remove data with zero value.}
}
\value{
A \code{tibble} object.
}
\description{
Get the layer group data in the form of a flat table that only includes the
geoid column of the geographic data.
}
\details{
Optionally you can delete the rows whose measurement value is zero.
}
\keyword{internal}
