% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_census.R
\name{get_file_year}
\alias{get_file_year}
\title{Get file year}
\usage{
get_file_year(file)
}
\arguments{
\item{name}{A string vector.}
}
\value{
A string vector.
}
\description{
Given a file, get the associated year (in the name).
}
\keyword{internal}
