% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{get_gbd_files}
\alias{get_gbd_files}
\title{Get gbd files}
\usage{
get_gbd_files(dir)
}
\arguments{
\item{name}{A string.}
}
\value{
A string vector.
}
\description{
Given a dir, we get the gbd files in it, at any level.
}
\keyword{internal}
