% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_5yr_topic.R
\name{get_geo_layer}
\alias{get_geo_layer}
\title{Get geo layer}
\usage{
get_geo_layer(file)
}
\arguments{
\item{file}{A string, file name.}
}
\value{
A \code{st}, geo data.
}
\description{
Get geo layer
}
\keyword{internal}
