% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_acs_5ye.R
\name{get_geodatabase_url}
\alias{get_geodatabase_url}
\title{get_geodatabase_url}
\usage{
get_geodatabase_url(url, extension, names, name, year)
}
\arguments{
\item{mdr}{A string.}
}
\value{
A boolean
}
\description{
get_geodatabase_url
}
\keyword{internal}
