% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_layer_get_layer.R
\name{get_layer_names}
\alias{get_layer_names}
\alias{get_layer_names.uscb_layer}
\title{get layer names}
\usage{
get_layer_names(ul)

\method{get_layer_names}{uscb_layer}(ul)
}
\arguments{
\item{ul}{A \code{uscb_layer} object.}
}
\value{
A vector of names.
}
\description{
get layer names.
}
\seealso{


Other data selection functions: 
\code{\link{uscb_layer}()}
}
\concept{data selection functions}
