% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_layer_get_layer.R
\name{get_layer_names}
\alias{get_layer_names}
\alias{get_layer_names.uscb_layer}
\title{Get layer names}
\usage{
get_layer_names(ul)

\method{get_layer_names}{uscb_layer}(ul)
}
\arguments{
\item{ul}{A \code{uscb_layer} object.}
}
\value{
A vector of names.
}
\description{
Once a specific geodatabase has been selected from which the class object has
been created, we can obtain the names of the layers it contains.
}
\examples{

folder <- system.file("extdata", package = "geogenr")
folder <- stringr::str_replace_all(paste(folder, "/", ""), " ", "")
ua <- uscb_acs_5ye(folder = folder)
sa <- ua |> get_statistical_areas()
# sa[6]
# [1] "New England City and Town Area Division"
ul <- uscb_layer(uscb_acs_metadata, ua = ua, geodatabase = sa[6], year = 2015)

layers <- ul |> get_layer_names()

}
\seealso{
Other data selection functions: 
\code{\link{get_layer_group_names}()},
\code{\link{get_layer_group}()},
\code{\link{get_layer}()},
\code{\link{uscb_folder}()},
\code{\link{uscb_layer}()}
}
\concept{data selection functions}
