% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_metadata_get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\alias{get_metadata.uscb_metadata}
\title{Get metadata}
\usage{
get_metadata(um)

\method{get_metadata}{uscb_metadata}(um)
}
\arguments{
\item{um}{A \code{uscb_metadata} object.}
}
\value{
A \code{uscb_metadata} object.
}
\description{
get the metadata to distribute it in columns according to its topic.
}
\keyword{internal}
