% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_5yr_topic.R
\name{get_names_of_other_topics}
\alias{get_names_of_other_topics}
\alias{get_names_of_other_topics.acs_5yr_topic}
\title{Get names of other topics (report groups)}
\usage{
get_names_of_other_topics(act)

\method{get_names_of_other_topics}{acs_5yr_topic}(act)
}
\arguments{
\item{act}{A \code{acs_5yr_topic} object.}
}
\value{
A vector, available topics.
}
\description{
The area that we have downloaded has a set of defined topics, we have selected
one of them, this function shows us the rest of the available topics in the
area.
}
\examples{

topics <- anrc_2021_x01 |>
  get_names_of_other_topics()

}
\seealso{
Other data selection functions: 
\code{\link{as_GeoPackage}()},
\code{\link{as_acs_5yr_geo}()},
\code{\link{as_acs_5yr_topic}()},
\code{\link{as_flat_table}()},
\code{\link{as_star_database}()},
\code{\link{get_available_area_topics}()},
\code{\link{get_available_area_years}()},
\code{\link{get_available_areas}()},
\code{\link{get_geo_attribute_names}()},
\code{\link{get_geo_layer.acs_5yr_geo}()},
\code{\link{get_metadata}()},
\code{\link{get_report_names}()},
\code{\link{get_subreport_names}()},
\code{\link{get_topic_name}()},
\code{\link{select_report}()},
\code{\link{select_subreport}()},
\code{\link{select_topic}()},
\code{\link{set_metadata}()}
}
\concept{data selection functions}
