% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_5yr_topic.R
\name{get_names_of_other_topics}
\alias{get_names_of_other_topics}
\alias{get_names_of_other_topics.acs_5yr_topic}
\title{Get names of other topics (report groups)}
\usage{
get_names_of_other_topics(act)

\method{get_names_of_other_topics}{acs_5yr_topic}(act)
}
\arguments{
\item{act}{A \code{acs_5yr_topic} object.}
}
\value{
A vector, available report groups.
}
\description{
Get the rest of the available topics with the one that is selected.
}
\examples{

dir <- tempdir()
source_dir <- system.file("extdata/acs_5yr", package = "geogenr")
files <- list.files(source_dir, "*.zip", full.names = TRUE)
file.copy(from = files, to = dir, overwrite = TRUE)
ac <- acs_5yr(dir)
files <- ac |>
  unzip_files()

act <- ac |>
  as_acs_5yr_topic("Alaska Native Regional Corporation",
                   topic = "X01 Age And Sex")

topics <- act |>
  get_names_of_other_topics()

}
\seealso{
Other data selection functions: 
\code{\link{as_acs_5yr_GeoPackage}()},
\code{\link{as_acs_5yr_topic}()},
\code{\link{as_flat_table}()},
\code{\link{as_star_database}()},
\code{\link{get_available_area_topics}()},
\code{\link{get_available_area_years}()},
\code{\link{get_available_areas}()},
\code{\link{get_report_names}()},
\code{\link{get_subreport_names}()},
\code{\link{get_topic_name}()},
\code{\link{select_report}()},
\code{\link{select_subreport}()},
\code{\link{select_topic}()}
}
\concept{data selection functions}
