% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_5yr_topic.R
\name{get_report_names}
\alias{get_report_names}
\alias{get_report_names.acs_5yr_topic}
\title{Get report names}
\usage{
get_report_names(act)

\method{get_report_names}{acs_5yr_topic}(act)
}
\arguments{
\item{act}{An \code{acs_5yr_topic} object.}
}
\value{
A vector, report names.
}
\description{
Each topic includes several reports. Once a topic has been selected, using this
function we obtain the name of the available reports. The report code is
included with the name. Each report can contain multiple subreports.
}
\examples{

reports <- anrc_2021_x01 |>
  get_report_names()

}
\seealso{
Other data selection functions: 
\code{\link{as_acs_5yr_topic}()},
\code{\link{get_available_area_topics}()},
\code{\link{get_available_area_years}()},
\code{\link{get_available_areas}()},
\code{\link{get_geo_attribute_names}()},
\code{\link{get_geo_layer.acs_5yr_geo}()},
\code{\link{get_names_of_other_topics}()},
\code{\link{get_subreport_names}()},
\code{\link{get_topic_name}()},
\code{\link{select_report}()},
\code{\link{select_subreport}()},
\code{\link{select_topic}()}
}
\concept{data selection functions}
