% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_census_download.R
\name{get_selected_file_names}
\alias{get_selected_file_names}
\alias{get_selected_file_names.acs_census}
\title{Get selected file names}
\usage{
get_selected_file_names(ac)

\method{get_selected_file_names}{acs_census}(ac)
}
\arguments{
\item{ac}{A \code{acs_census} object.}
}
\value{
A vector, file names.
}
\description{
Gets the names of the files selected to be downloaded.
}
\examples{

dir <- system.file("extdata", package = "geogenr")
ac <- acs_census(dir)

groups <- ac |>
  get_selected_file_names()

}
\seealso{
Other data download functions: 
\code{\link{acs_census}()},
\code{\link{download_selected_files}()},
\code{\link{get_area_file_names}()},
\code{\link{get_area_groups}()},
\code{\link{get_area_years}()},
\code{\link{get_areas}()},
\code{\link{get_too_heavy_file_names}()},
\code{\link{select_area_files}()},
\code{\link{unzip_files}()}
}
\concept{data download functions}
