% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_metadata.R
\name{\%>\%}
\alias{\%>\%}
\alias{new_uscb_metadata}
\title{\code{uscb_metadata} S3 class}
\usage{
new_uscb_metadata(filepath = NULL, code = NULL, group_code = NULL)
}
\arguments{
\item{filepath}{A string, path to gbd file.}

\item{code}{A string.}
}
\value{
A \code{uscb_metadata} object.
}
\description{
Internal low-level constructor that creates new objects with the correct
structure.
}
\keyword{internal}
