% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_acs_5ye.R, R/uscb_folder.R,
%   R/uscb_layer.R, R/uscb_metadata.R
\name{\%>\%}
\alias{\%>\%}
\alias{new_uscb_acs_5ye}
\alias{new_uscb_folder}
\alias{new_uscb_layer}
\alias{new_uscb_metadata}
\title{\code{uscb_acs_5ye} S3 class}
\usage{
new_uscb_acs_5ye(folder = "")

new_uscb_folder(ul, year = NULL)

new_uscb_layer(metadata, ua, geodatabase, year, folder = NULL)

new_uscb_metadata(
  filepath = NULL,
  code = NULL,
  group_code = NULL,
  short_name = NULL,
  uscb_acs_metadata = NULL
)
}
\arguments{
\item{folder}{A string.}

\item{ul}{A \code{uscb_folder} object.}

\item{year}{A integer}

\item{metadata}{A metadata object.}

\item{ua}{A \code{uscb_acs_5ye} object.}

\item{geodatabase}{A string.}

\item{filepath}{A string, path to gbd file.}

\item{code}{A vector of strings.}

\item{group_code}{A vector of strings.}

\item{short_name}{A vector of strings.}

\item{uscb_acs_metadata}{A metadata object.}
}
\value{
A \code{uscb_acs_5ye} object.

A \code{uscb_folder} object.

A \code{uscb_layer} object.

A \code{uscb_metadata} object.
}
\description{
Internal low-level constructor that creates new objects with the correct
structure.

Internal low-level constructor that creates new objects with the correct
structure.

Internal low-level constructor that creates new objects with the correct
structure.

Internal low-level constructor that creates new objects with the correct
structure.
}
\keyword{internal}
