% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_metadata_interpret_metadata.R
\name{interpret_metadata}
\alias{interpret_metadata}
\alias{interpret_metadata.uscb_metadata}
\title{Interpret metadata}
\usage{
interpret_metadata(um)

\method{interpret_metadata}{uscb_metadata}(um)
}
\arguments{
\item{um}{A \code{uscb_metadata} object.}
}
\value{
A \code{uscb_metadata} object.
}
\description{
Interpret the metadata to distribute it in columns according to its topic.
}
\keyword{internal}
