% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_census_download.R
\name{select_area_files}
\alias{select_area_files}
\alias{select_area_files.acs_census}
\title{Select area files}
\usage{
select_area_files(ac, area, years)

\method{select_area_files}{acs_census}(ac, area, years = NULL)
}
\arguments{
\item{ac}{A \code{acs_census} object.}

\item{area}{A string, area name.}

\item{years}{A vector, year number.}
}
\value{
A \code{acs_census} object.
}
\description{
Select area files for the given years. If no year is indicated, all available
ones are selected.
}
\examples{

dir <- system.file("extdata", package = "geogenr")
ac <- acs_census(dir)

ac <- ac |>
  select_area_files("State", 2019:2021)

ac <- ac |>
  select_area_files("State")

}
\seealso{
Other data download functions: 
\code{\link{acs_census}()},
\code{\link{download_selected_files}()},
\code{\link{get_area_file_names}()},
\code{\link{get_area_groups}()},
\code{\link{get_area_years}()},
\code{\link{get_areas}()},
\code{\link{get_selected_file_names}()},
\code{\link{get_too_heavy_file_names}()},
\code{\link{unzip_files}()}
}
\concept{data download functions}
