% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs_5yr_topic.R
\name{select_report}
\alias{select_report}
\alias{select_report.acs_5yr_topic}
\title{Select report}
\usage{
select_report(act, report)

\method{select_report}{acs_5yr_topic}(act, report = NULL)
}
\arguments{
\item{act}{A \code{acs_5yr_topic} object.}
}
\value{
A vector, topic name.
}
\description{
Get the selected topic.
}
\examples{

reports <- anrc_2021_x01 |>
  select_report()

}
\seealso{
Other data selection functions: 
\code{\link{as_GeoPackage}()},
\code{\link{as_acs_5yr_geo}()},
\code{\link{as_acs_5yr_topic}()},
\code{\link{as_flat_table}()},
\code{\link{as_star_database}()},
\code{\link{get_available_area_topics}()},
\code{\link{get_available_area_years}()},
\code{\link{get_available_areas}()},
\code{\link{get_geo_attribute_names}()},
\code{\link{get_geo_layer.acs_5yr_geo}()},
\code{\link{get_metadata}()},
\code{\link{get_names_of_other_topics}()},
\code{\link{get_report_names}()},
\code{\link{get_subreport_names}()},
\code{\link{get_topic_name}()},
\code{\link{select_subreport}()},
\code{\link{select_topic}()},
\code{\link{set_metadata}()}
}
\concept{data selection functions}
