% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_obtaining.R
\name{unzip_area_file}
\alias{unzip_area_file}
\title{Unzip area file}
\usage{
unzip_area_file(
  file,
  out_dir = NULL,
  delete_zip = FALSE,
  only_show_files = FALSE
)
}
\arguments{
\item{file}{A string or string vector.}

\item{out_dir}{A string or string vector, output folder.}

\item{delete_zip}{A boolean, delete zip files if correctly unzipped.}

\item{only_show_files}{A boolean, only show the files that would be unzipped,
and the destination folders, not unzip them.}
}
\value{
A vector of strings, name of the processed files.
}
\description{
Given a vector of compressed file names or the name of a folder containing
compressed files, unzip the files to the given output folder.
}
\examples{

dir <- tempdir()
url <- get_area_file_names("State", 2021)

url <-
  paste0(
    'file://',
    system.file("extdata/ACS_2014_5YR_NECTADIV.gdb.zip", package = "geogenr")
  )

files <- download_area_file(url, dir, unzip = FALSE)

res <- unzip_area_file(files, tempdir())

}
\seealso{
Other data download functions: 
\code{\link{download_area_file}()},
\code{\link{get_area_file_names}()},
\code{\link{get_area_groups}()},
\code{\link{get_area_years}()},
\code{\link{get_areas}()}
}
\concept{data download functions}
