% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_acs_5ye.R
\name{uscb_acs_5ye}
\alias{uscb_acs_5ye}
\title{\code{uscb_acs_5ye} S3 class}
\usage{
uscb_acs_5ye(folder = "")
}
\arguments{
\item{folder}{A string.}
}
\value{
A \code{uscb_acs_5ye} object.
}
\description{
A \code{uscb_acs_5ye} object is created from a given local folder.
}
\examples{

folder <- "../geodimension/data/us/"
ua <- uscb_acs_5ye(folder = folder)

folder <- system.file("extdata", package = "geogenr")
folder <- stringr::str_replace_all(paste(folder, "/", ""), " ", "")
ua <- uscb_acs_5ye(folder = folder)

}
\seealso{
Other data collection functions: 
\code{\link{download_geodatabases}()},
\code{\link{get_available_years_downloaded}()},
\code{\link{get_available_years_in_the_web}()},
\code{\link{get_legal_and_administrative_areas}()},
\code{\link{get_statistical_areas}()}
}
\concept{data collection functions}
