% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_layer.R
\name{uscb_layer}
\alias{uscb_layer}
\title{\code{uscb_layer} S3 class}
\usage{
uscb_layer(metadata, ua, geodatabase, year, folder = NULL)
}
\arguments{
\item{metadata}{A metadata object.}

\item{ua}{A \code{uscb_acs_5ye} object.}

\item{geodatabase}{A string.}

\item{year}{A integer}

\item{folder}{A string.}
}
\value{
A \code{uscb_layer} object.
}
\description{
A \code{uscb_layer} object is created from a given
}
