% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uscb_metadata.R
\name{uscb_metadata}
\alias{uscb_metadata}
\title{\code{uscb_metadata} S3 class}
\usage{
uscb_metadata(
  filepath = NULL,
  code = NULL,
  group_code = NULL,
  short_name = NULL,
  uscb_acs_metadata = NULL
)
}
\arguments{
\item{filepath}{A string, path to gbd file.}

\item{code}{A vector of strings.}

\item{group_code}{A vector of strings.}

\item{short_name}{A vector of strings.}

\item{uscb_acs_metadata}{A metadata object.}
}
\value{
A \code{uscb_metadata} object.
}
\description{
A \code{uscb_metadata} object is created from a given geodatabase, code, group
code, short name and existing metadata.
}
\keyword{internal}
